refactor: extract abstract base-class
We will extend from this baseclass with our new SVG class.
